fix windows build
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@1833 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/applications/mod_conference/mod_conference.c
@@ -557,7 +557,7 @@ static void *SWITCH_THREAD_FUNC conference_thread_run(switch_thread_t *thread, v
 						for (x = 0; x < imember->read / 2; x++) {
 							int32_t z = muxed[x] + bptr[x];
 							normalize_to_16bit(z);
-							muxed[x] = z;
+							muxed[x] = (int16_t)z;
 						}
 						
 						ready++;
